requirements.txt
pandas==0.25.3
psycopg2-binary==2.8.4
